Разбил классы на разные файлы, добавил проверки на пустой ввод, символьный ввод, проверка на одинаковую скорость у всех машин
