+++ src/main/java/Car.java
@@ -1,0 +1,10 @@
+public class Car {
+
+    String name;
+    int speed;
+
+    public Car(String name, int speed) {
+        this.name = name;
+        this.speed = speed;
+    }
+}
